Move the Connection to Connection\Handler
There is a lot of ambiguity between the Hoa\Socket objects and Hoa\Websocket extension.

I've moved the Handler part inside an Handler object
